OPENSC belongs in profile to execute just once

+++ .bash_profile
@@ -11,13 +11,28 @@ addPath -"$HOME/bin"
 export PATH MANPATH
 
 if [[ "$-" == *i* ]] || tty -s ; then
-  which ssh-agent &>/dev/null && {
+  if which ssh-agent &>/dev/null; then
     if [ -z "$SSH_AUTH_SOCK" ]; then
         eval `ssh-agent ${SSH_AGENT_ARGS:-${BASH_VERSION:+ -s}}`
         trap "kill -9 $SSH_AGENT_PID" EXIT
     fi
     [ -n "$SSH_AUTH_SOCK" ] && ssh-add -k 2>/dev/null
-  }
+
+    # CAC/PIF card support
+    case ${OSTYPE:-`uname`} in
+      [lL]inux*)
+            : ${OPENSC_LIB:=/usr/lib/`uname -m`-linux-gnu/opensc-pkcs11.so} ;;
+      cygwin|CYGWIN_*)
+            # MSI doesn't allow tailoring path
+            : ${OPENSC_LIB:="`convert_path -p "$PROGRAMFILES"`/OpenSC Project/OpenSC/pkcs11/opensc-pkcs11.dll"} ;;
+      [dD]arwin*)
+            : ${OPENSC_LIB:=/usr/local/lib/opensc-pkcs11.so}
+    esac
+    if [ -f "$OPENSC_LIB" ]; then
+      export OPENSC_LIB
+      [ -n "$SSH_AUTH_SOCK" ] && ssh-add -s "$OPENSC_LIB" 2>/dev/null
+    fi
+  fi
 fi
 
 for f in "$HOME"/.{bash_profile.local,bashrc.${OSTYPE:=`uname`}}; do
@@ -31,4 +46,7 @@ done
 export EDITOR PAGER
 source "$HOME"/.bashrc
 
+
+
+return 0
 # vim: set expandtab:ts=4:sw=4
